feat: parse reported and item elements in form
