+++ src/connection/FormItemField.ts
@@ -1,0 +1,5 @@
+import FormField from "./FormField";
+
+export default class FormItemField extends FormField {
+
+}
