+++ src/connection/FormReportedField.ts
@@ -1,0 +1,5 @@
+import FormField from "./FormField";
+
+export default class FormReportedField extends FormField {
+
+}
