+++ src/errors/ParsingError.ts
@@ -1,0 +1,5 @@
+import BaseError from './BaseError'
+
+export default class ParsingError extends BaseError {
+
+}
